# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: neural-link
    description: Profile-grounded AI conversations and voice
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /chat:
    post:
      operationId: sendChatMessage
      tags: [neural-link]
      summary: Ask the Neural Link assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid message
        "502":
          description: AI provider unavailable
  /tts:
    post:
      operationId: synthesizeSpeech
      tags: [neural-link]
      summary: Generate spoken assistant audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsInput"
      responses:
        "200":
          description: MP3 audio
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid text
        "502":
          description: AI provider unavailable
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChatInput:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 24
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - messages
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - role
        - content
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
        grounded:
          type: boolean
      required:
        - answer
        - grounded
    TtsInput:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - text