# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  /**
     * @minItems 1
     * @maxItems 24
     */
  messages: ChatMessage[];
}