# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  role: ChatMessageRole;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}